fixed theme uninstallation

+++ Rectify11Installer/Core/Rectify11Installer.cs
@@ -291,7 +291,7 @@ namespace Rectify11Installer
                     {
                         Directory.Delete(@"C:\Windows\Resources\themes\rectify11");
                     }
-                    if (File.Exists(@"C:\Windows\Resources\theme\lightrectified.theme") || File.Exists(@"C:\Windows\Resources\theme\darkrectified.theme") || File.Exists(@"C:\Windows\Resources\theme\darkcolorized.theme") || File.Exists(@"C:\Windows\Resources\theme\black.theme") || File.Exists(@"C:\Windows\Resources\theme\blacknonhighcontrastribbon.theme"))
+                    if (File.Exists(@"C:\Windows\Resources\theme\lightrectified.theme") && File.Exists(@"C:\Windows\Resources\theme\darkrectified.theme") && File.Exists(@"C:\Windows\Resources\theme\darkcolorized.theme") && File.Exists(@"C:\Windows\Resources\theme\black.theme") && File.Exists(@"C:\Windows\Resources\theme\blacknonhighcontrastribbon.theme"))
                     {
                         File.Delete(@"C:\Windows\Resources\theme\lightrectified.theme");
                         File.Delete(@"C:\Windows\Resources\theme\darkrectified.theme");
